requirements.txt
Flask==1.1.2
flask-restplus==0.13.0
numpy==1.19.4
python-dotenv==0.15.0
Werkzeug==0.16.1